Rename demos api router file
