+++ apps/api-hono/src/app.ts
@@ -1,7 +1,7 @@
 import { Hono } from "hono";
 import { logger } from "hono/logger";
 
-import demoApi from "#routes/demo.ts";
+import demoApi from "#routes/demos.ts";
 import projectApi from "#routes/projects.ts";
 import { notFound, onError } from "#utils/route.utils.ts";
